requirements.txt
Flask>=2.0
PyPDF2>=3.0
nltk>=3.6
google-generativeai>=0.2.0
python-dotenv>=1.0
